// movie-app/src/helpers/type.ts
export type SearchOptionType = {
  value: string;
  label: string;
};

export type SearchItemType = {
  search: string;
  year: string;
  type: string;
};
export type MovieType = {
  Poster: string;
  Title: string;
  Type: string;
  Year: string;
  imdbID: string;
};

export type DetailMovieType = {
  Poster: string;
  Title: string;
  Type: string;
  Year: string;
  Actors: string;
  Awards: string;
  BoxOffice: string;
  Country: string;
  DVD: string;
  Director: string;
  Genre: string;
  Language: string;
  Metascore: string;
  Plot: string;
  Production: string;
  Rated: string;
  Released: string;
  imdbRating: string;
  imdbVotes: string;
  imdbID: string;
  Writer: string;
};